+++ changes.md
@@ -3,7 +3,9 @@
 #### ?
 -   Added: `NativeEmailHandler` now available for logging
 -   Added: Details view for log entries
+-   Added: Priority for jobs, can be run on a high or low priority queue handler
 -   Changed: Switched the logs table to MyISAM engine
+-   Changed: Jobs are running on the laravel queue
 
 #### [1.0.2] 2015-10-26
 -   Added: `ConnectorLoggerInterface` is now available in `Stage` and `CronHandler` class as $this->logger